+++ src/components/sidebar/right-sidebar/datepicker/Datepicker.tsx
@@ -8,33 +8,33 @@ import { updateFilters } from "../../../../utils/stateUpdateUtils";
 import Translate from "../../../../utils/translate/translateService";
 import SectionHeader from "../SectionHeader";
 import {enUS, fr} from 'date-fns/locale'
-import DateSlider from "./DateSlider";
-import { sendFiltersAnalyticsEvent } from "../../../../utils/analyticsUtils";
-import NewDateSlider from '../new-datepicker/NewDateSlider';
+import DateSlider from './DateSlider';
+import '../datepicker/DateSlider.css'
 
 interface DatepickerProps {
   page: string
 }
 
 export default function Datepicker({ page }: DatepickerProps) {
+
   const [state, dispatch] = useContext(AppContext);
   const filters = (state[page as keyof State] as PageState).filters;
-  const [filterStartDate, filterEndDate]: Date[] = Array.from(filters.publish_date);
-  const [chosenStartDate, setChosenStartDate] = useState<Date>(filterStartDate);
-  const [chosenEndDate, setChosenEndDate] = useState<Date>(filterEndDate);
   const earliestPublicationDate = state.calendarStartDates.minDate;
   const latestPublicationDate = state.calendarStartDates.maxDate;
   const millisecondsPerDay = 86400000;
   const [sliderRange, setSliderRange] = useState<number>((latestPublicationDate.getTime()-earliestPublicationDate.getTime())/millisecondsPerDay); //in days
-  const [sliderThumbValues, setSliderThumbValues] = useState<Date[]>([chosenStartDate, chosenEndDate]);
+  const [filterStartDate, filterEndDate]: Date[] = Array.from(filters.publish_date);
 
-  useEffect(() => {
-    setSliderRange((latestPublicationDate.getTime()-earliestPublicationDate.getTime())/millisecondsPerDay);
-    //although it is not possible for a suer to cause the below problem that I am checking for,
-    // there does come an issue that if a user waits for the whole page to load, then the chosen date is earlier than the earliest possible date
-    setChosenStartDate(earliestPublicationDate);
-    setChosenEndDate(latestPublicationDate);
-  }, [latestPublicationDate, earliestPublicationDate])
+  // TODO: should these functions be in util since one of them is used in both this file and the slider file?
+  const fromDateToNumber = (date: Date) => {
+    return (date.getTime()-earliestPublicationDate.getTime())/millisecondsPerDay
+  }
+  const toDateSinceMinDate = (val: number) => {
+    return new Date(earliestPublicationDate.getTime() + (val*millisecondsPerDay));
+  }
+
+  //dates as number of days since mindate
+  const [chosenDates, setChosenDates] = useState<number[]>([fromDateToNumber(filterStartDate), fromDateToNumber(filterEndDate)])
 
   useEffect(() => {
     registerLocale("en", enUS)
@@ -42,97 +42,97 @@ export default function Datepicker({ page }: DatepickerProps) {
   }, [])
 
   useEffect(() => {
-    setSliderThumbValues([chosenStartDate, chosenEndDate]);
-  }, [chosenStartDate, chosenEndDate])
+    setSliderRange(fromDateToNumber(latestPublicationDate));
+    }, [latestPublicationDate, earliestPublicationDate])
 
-  const datePickerChanged = (isStart: Boolean, date: Date) => {
-    const newDates = [chosenStartDate, chosenEndDate];
-    newDates[0] = isStart ? date : chosenStartDate;
-    newDates[1] = !isStart ? date : chosenEndDate;
+  const onChange = (dates: number[]) => {
+      setChosenDates(dates)
+  }
+
+  const onUpdate = (isStart: Boolean, date: Date) => {
+    let newDates: Date[];
 
     if (isStart) {
-      setChosenStartDate(date);
+      newDates = [date, toDateSinceMinDate(chosenDates[1])];
+      setChosenDates([fromDateToNumber(date), chosenDates[1]])
     }
     else {
-      setChosenEndDate(date);
+      newDates = [toDateSinceMinDate(chosenDates[0]), date];
+      setChosenDates([chosenDates[0], fromDateToNumber(date)])
     }
+
     updateFilters(dispatch,
-      filters,
-      "publish_date" as FilterType,
-      newDates,
-      page)
+        filters,
+        "publish_date" as FilterType,
+        newDates,
+        page)
   }
 
   const CustomInput = ({ value, onClick, text }: any) => (
-    <div className="col-12 p-0 flex column">
-      <div>
-        {text}
+      <div className="col-12 p-0 flex input-seperate">
+        <Input className="col-12 p-0 cursor"
+               value={value}
+               fluid={true}
+               type="text"
+               // TODO: also add translation in german.json
+               placeholder={Translate('SelectDate')}
+               onClick={onClick} />
       </div>
-      <Input className="col-12 p-0 cursor"
-        value={value}
-        iconPosition="left"
-        icon="calendar"
-        fluid={true}
-        type="text"
-        placeholder='Select date..'
-        onClick={onClick} />
-    </div>
   )
 
   return (
-    <div className="row justify-content-center">
-      <div className="col-10 col align-items-center p-0">
-        <div className="pb-1">
-          <div>
-            <SectionHeader header_text={Translate('DateRange')} tooltip_text={Translate('DateRangeTooltip')} />
-          </div>
-          <DateSlider maxPossibleValue={sliderRange} minPossibleValue={0} onMouseUp={datePickerChanged} values={sliderThumbValues} minDate={earliestPublicationDate}/>
-          <div>
-            <DatePicker
-                selected={chosenStartDate}
-                onChange={() => { }}
-                onSelect={(date: Date) => datePickerChanged(true, date)}
-                dateFormatCalendar={"MMMM yyyy "}
-                dateFormat="yyyy/MM/dd"
-                minDate={earliestPublicationDate}
-                maxDate={chosenEndDate}
-                showYearDropdown
-                yearDropdownItemNumber={5}
-                scrollableYearDropdown
-                showMonthDropdown
-                locale={state.language}
-                customInput={<CustomInput value={chosenStartDate} text={Translate("StartDate")} onClick={onclick} />}
-                withPortal
-                closeOnScroll={true}
-                shouldCloseOnSelect={false}
-                dropdownMode="select"
-                todayButton="Today"
-            />
-          </div>
-          <div>
-            <DatePicker
-                selected={chosenEndDate}
-                onChange={() => { }}
-                onSelect={(date: Date) => datePickerChanged(false, date)}
-                dateFormat="yyyy/MM/dd"
-                dateFormatCalendar={"MMM yyyy"}
-                minDate={chosenStartDate}
-                maxDate={latestPublicationDate}
-                showYearDropdown
-                yearDropdownItemNumber={5}
-                scrollableYearDropdown
-                showMonthDropdown
-                locale={state.language}
-                customInput={<CustomInput value={chosenEndDate} text={Translate("EndDate")} onClick={onclick} />}
-                withPortal
-                closeOnScroll={true}
-                shouldCloseOnSelect={false}
-                dropdownMode="select"
-                todayButton="Today"
-            />
+      <div className="row justify-content-center">
+        <div className="col-10 col align-items-center p-0">
+          <div className="pb-1">
+            <div>
+              <SectionHeader header_text={Translate('DateRange')} tooltip_text={Translate('DateRangeTooltip')} />
+            </div>
+            <DateSlider maxPossibleValue={sliderRange} minPossibleValue={0} onMouseUp={onUpdate} onSliderMove={onChange} values={chosenDates} minDate={earliestPublicationDate}/>
+            <div className="space-between mt-3 mb-3">
+              <DatePicker
+                  selected={toDateSinceMinDate(chosenDates[0])}
+                  onChange={() => { }}
+                  onSelect={(date: Date) => onUpdate(true, date)}
+                  dateFormatCalendar={"MMMM yyyy "}
+                  dateFormat="yyyy/MM/dd"
+                  minDate={earliestPublicationDate}
+                  maxDate={toDateSinceMinDate(chosenDates[1])}
+                  showYearDropdown
+                  yearDropdownItemNumber={5}
+                  scrollableYearDropdown
+                  showMonthDropdown
+                  locale={state.language}
+                  customInput={<CustomInput value={toDateSinceMinDate(chosenDates[0])} onClick={onclick} />}
+                  withPortal
+                  closeOnScroll={true}
+                  shouldCloseOnSelect={false}
+                  dropdownMode="select"
+                  // TODO: also add translation in german.json
+                  todayButton={Translate("Today")}
+              />
+               <DatePicker
+                  selected={toDateSinceMinDate(chosenDates[1])}
+                  onChange={() => { }}
+                  onSelect={(date: Date) => onUpdate(false, date)}
+                  dateFormat="yyyy/MM/dd"
+                  dateFormatCalendar={"MMM yyyy"}
+                  minDate={toDateSinceMinDate(chosenDates[0])}
+                  maxDate={latestPublicationDate}
+                  showYearDropdown
+                  yearDropdownItemNumber={5}
+                  scrollableYearDropdown
+                  showMonthDropdown
+                  locale={state.language}
+                  customInput={<CustomInput value={toDateSinceMinDate(chosenDates[1])} onClick={onclick} />}
+                  withPortal
+                  closeOnScroll={true}
+                  shouldCloseOnSelect={false}
+                  dropdownMode="select"
+                  todayButton="Today"
+              />
+            </div>
           </div>
         </div>
       </div>
-    </div>
   )
 }
